added edit button for an admin

+++ app/auth/router.py
@@ -7,6 +7,8 @@ from sqlalchemy.exc import IntegrityError
 from app.auth.schema import *
 from bot import bot
 from aiogram import types
+from aiogram.types.web_app_info import WebAppInfo
+from telegram.superuser.inlinekeyboards import get_buttons_for_new_admin
 
 router = APIRouter(
     prefix='/api/v1',
@@ -45,6 +47,7 @@ async def register(background_tasks: BackgroundTasks,admin: AdminCreateSchema, c
 
         # Форматирование объекта datetime в нужный формат
         formatted_date = db_admin.created_at.strftime("%d %B %Y %H:%M")
+        
         message_text = (
             "Добавлен новый Админ\n\n"
             f'Номер Админа: {db_admin.id}\n'
@@ -55,13 +58,11 @@ async def register(background_tasks: BackgroundTasks,admin: AdminCreateSchema, c
             f'Фамилия: {db_admin.last_name}\n'
             f'Номер телефона: {db_admin.phone_number}\n'
             f'ID канала: {db_admin.channel_id}\n'
+            f'Супер Админ: {db_admin.is_superuser}\n'
             f"Дата создания: {formatted_date}"
         )
-        buttons_new_admin = types.InlineKeyboardMarkup()
-        delete_button = types.InlineKeyboardButton("Удалить", callback_data=f"delete_created_admin:{db_admin.id}")
-        close_msg_button = types.InlineKeyboardButton("Скрыть", callback_data=f"close_msg")
-        buttons_new_admin.add(delete_button).add(close_msg_button)
-        await bot.send_message(chat_id=current_user.tg_id, text=message_text, reply_markup=buttons_new_admin)
+        btns = get_buttons_for_new_admin(db_admin)
+        await bot.send_message(chat_id=current_user.tg_id, text=message_text, reply_markup=btns)
 
     background_tasks.add_task(send_message_task)
 
@@ -90,6 +91,36 @@ async def get_own_info(current_user=Depends(get_current_user)):
 
 @router.delete('/delete/admin/{admin_id}')
 async def delete_admin(admin_id, current_user=Depends(get_current_user), db: Session = Depends(get_db)):
+    if current_user.is_superuser == False:
+        raise HTTPException(status_code=403, detail="Недостаточно прав доступа.")
+    
     db_admin = db.query(models.Admin).filter(models.Admin.id == admin_id).first()
     db.delete(db_admin)
     db.commit()
+
+@router.put('/admin/{admin_id}', name="update admin data", response_model=AdminSchema)
+async def update_admin_data(admin_id: int, admin_update: AdminUpdateSchema, current_user=Depends(get_current_user), db: Session = Depends(get_db)):
+    if current_user.is_superuser == False:
+        raise HTTPException(status_code=403, detail="Недостаточно прав доступа.")
+    
+    db_admin = db.query(models.Admin).filter(models.Admin.id == admin_id).first()
+    if admin_update.tg_id is not None:
+        db_admin.tg_id = admin_update.tg_id
+    if admin_update.email is not None:
+        db_admin.email = admin_update.email
+    if admin_update.username is not None:
+        db_admin.username = admin_update.username
+    if admin_update.first_name is not None:
+        db_admin.first_name = admin_update.first_name
+    if admin_update.last_name is not None:
+        db_admin.last_name = admin_update.last_name
+    if admin_update.phone_number is not None:
+        db_admin.phone_number = admin_update.phone_number
+    if admin_update.channel_id is not None:
+        db_admin.channel_id = admin_update.channel_id
+    if admin_update.is_superuser is not False:
+        db_admin.is_superuser = admin_update.is_superuser
+
+    db.commit()
+
+    return db_admin
